% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_branch.R
\name{add_filter_branch}
\alias{add_filter_branch}
\alias{add_filter_branch.mverse}
\title{Add variable branches to a \code{mverse} object.}
\usage{
add_filter_branch(.mverse, ...)

\method{add_filter_branch}{mverse}(.mverse, ...)
}
\arguments{
\item{.mverse}{a \code{mverse} object.}

\item{...}{\code{filter_branch} objects.}
}
\value{
The resulting \code{mverse} object.
}
\description{
This method adds one or more filter branches to
an existing \code{mverse} object.
}
\examples{
\dontrun{
mv <- create_multiverse(df)
x3 <- filter_branch(x1+x2, mean(c(x1,x2)))
mv <- mv \%>\%
  add_filter_branch(x3)
}
}
\seealso{
Other {methods for working with a filter branch}: 
\code{\link{filter_branch}()}
}
\concept{{methods for working with a filter branch}}
