% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_branch.R
\name{add_formula_branch}
\alias{add_formula_branch}
\alias{add_formula_branch.mverse}
\title{Add formula branches to a \code{mverse} object.}
\usage{
add_formula_branch(.mverse, ...)

\method{add_formula_branch}{mverse}(.mverse, ...)
}
\arguments{
\item{.mverse}{a \code{mverse} object.}

\item{...}{\code{formula_branch} objects.}
}
\value{
The resulting \code{mverse} object.
}
\description{
This method adds one or more formula branches to
an existing \code{mverse} object. The formula branches
are used to specify a model for the main analysis.
}
\examples{
\dontrun{
mv <- create_multiverse(df)
model_spec <- formula_branch(y ~ x1 + x2, y ~ x1 * x2)
mv <- mv \%>\%
  add_formula_branch(model_spec)
}
}
\seealso{
Other {methods for working with a formula branch}: 
\code{\link{family_branch}()},
\code{\link{formula_branch}()}
}
\concept{{methods for working with a formula branch}}
