% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_branch.R
\name{add_mutate_branch}
\alias{add_mutate_branch}
\alias{add_mutate_branch.mverse}
\title{Add variable branches to a \code{mverse} object.}
\usage{
add_mutate_branch(.mverse, ...)

\method{add_mutate_branch}{mverse}(.mverse, ...)
}
\arguments{
\item{.mverse}{a \code{mverse} object.}

\item{...}{\code{mutate_branch} objects.}
}
\value{
The resulting \code{mverse} object.
}
\description{
This method adds one or more mutate branches to
an existing \code{mverse} object.
}
\examples{
\dontrun{
mv <- create_multiverse(df)
x3 <- mutate_branch(x1+x2, mean(c(x1,x2)))
mv <- mv \%>\%
  add_mutate_branch(x3)
}
}
\seealso{
Other {methods for working with a mutate branch}: 
\code{\link{mutate_branch}()}
}
\concept{{methods for working with a mutate branch}}
