% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{add_variable_branch}
\alias{add_variable_branch}
\title{Add variable branch(es) to a \code{multiverse}.}
\usage{
add_variable_branch(.multiverse)
}
\arguments{
\item{.multiverse}{A \code{multiverse} object.}

\item{branch}{A variable branch rule.}
}
\description{
This function adds one ore more variable branch to
an existing \code{multiverse} object.
}
