% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition.R
\name{branch_condition}
\alias{branch_condition}
\title{Create a new branch condition.}
\usage{
branch_condition(x, y, reject = FALSE)
}
\arguments{
\item{x}{option 1}

\item{y}{option 2}

\item{reject}{if TRUE, the condition rejects universes
with option 1 and option 2}
}
\value{
A \code{branch_condition} object.
}
\description{
See [add_branch_condition] for examples.
}
\seealso{
Other {methods for working with a branch condition}: 
\code{\link{add_branch_condition}()}
}
\concept{{methods for working with a branch condition}}
