% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{coxph_mverse}
\alias{coxph_mverse}
\title{Fit \code{coxph} across the multiverse.}
\usage{
coxph_mverse(.mverse, ties = NULL)
}
\arguments{
\item{.mverse}{a \code{mverse} object.}
}
\value{
A \code{mverse} object with \code{coxph} fitted.
}
\description{
\code{coxph_mverse} fits \code{coxph} across the multiverse
according to model specifications provided by \code{formula_branch}.
}
\seealso{
Other {model fitting methods}: 
\code{\link{lm_mverse}()}
}
\concept{{model fitting methods}}
