% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{create_multiverse}
\alias{create_multiverse}
\title{Define a \code{multiverse} with data.}
\usage{
create_multiverse(data)
}
\arguments{
\item{data}{A raw data set to create a multiverse from.}
}
\value{
A \code{multivese} object with the source data
  stored in \code{$source}.
}
\description{
This function defines a \code{multiverse} object with
the given data set.
}
\examples{
mverse <- create_multiverse(matrix(c(1:9), nrow=3))
print(mverse$source)
}
