% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mverse.R
\name{extract}
\alias{extract}
\alias{extract.mverse}
\title{Extract branched values.}
\usage{
extract(...)

\method{extract}{mverse}(.mverse, columns = NULL, universe = 0, how = "all", n = 0)
}
\arguments{
\item{.mverse}{A \code{mverse} object.}

\item{columns}{(optional) A vector of characters.
Default: NULL.}

\item{universe}{(optional) Integer.
Default: 0.}

\item{how}{(optional) 'all', 'manual', or 'random'.
Default: 'all'.}

\item{n}{(optional) Integer.
Default 0.}
}
\description{
This method returns a table of selected values
across the multiverse in a long format. You
can select a subset based on the universe and
columns.
}
