% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family_branch.R
\name{family_branch}
\alias{family_branch}
\title{Create a new family branch.}
\usage{
family_branch(..., name = NULL)
}
\arguments{
\item{...}{branch definition expressions.}

\item{name}{Name for the new family.}
}
\value{
a \code{family_branch} object.
}
\description{
Create a new family branch.
}
\examples{
\dontrun{
fam <- family_branch(
  poisson, gaussian(link = "log")
)
}

}
\seealso{
Other {methods for working with a formula branch}: 
\code{\link{add_formula_branch}()},
\code{\link{formula_branch}()}
}
\concept{{methods for working with a formula branch}}
