% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_branch.R
\name{filter_branch}
\alias{filter_branch}
\title{Create a new filter branch.}
\usage{
filter_branch(..., name = NULL)
}
\arguments{
\item{...}{branch definition expressions.}

\item{name}{Name for the new filter.}
}
\value{
a filter_branch object.
}
\description{
Create a new filter branch.
}
\examples{
\dontrun{
x4 <- filter_branch(x1 + x2, x1 + x3, sum(x1, x2, x3))
}

}
\seealso{
Other {methods for working with a filter branch}: 
\code{\link{add_filter_branch}()}
}
\concept{{methods for working with a filter branch}}
