% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{fit_glm}
\alias{fit_glm}
\alias{glm_mverse}
\title{Fit \code{glm} across the multiverse.}
\usage{
glm_mverse(.mverse, family)
}
\arguments{
\item{.mverse}{a \code{mverse} object.}

\item{family}{a description of the error distribution and link function to be used in the model.}
}
\value{
A \code{mverse} object with \code{glm} fitted.
}
\description{
\code{glm_mverse} fits \code{glm} across the multiverse
according to model specifications provided by \code{formula_branch}.
At least one \code{formula_branch} must have been added.
}
\examples{
\dontrun{
model_spec <- formula_branch(y ~ x1)
mv <- mv \%>\%
  add_formula_branch(model_spec) \%>\%
  fit_glm()
}
}
\seealso{
Other {model fitting methods}: 
\code{\link{fit_lm}}
}
\concept{{model fitting methods}}
