% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htest.R, R/model.R
\name{fit_lm}
\alias{fit_lm}
\alias{ttest_mverse}
\alias{lm_mverse}
\title{Performs one and two sample t-tests on data columns.}
\usage{
ttest_mverse(
  .mverse,
  x,
  y = NULL,
  alternative = "two.sided",
  mu = 0,
  paired = FALSE,
  var.equal = FALSE,
  conf.level = 0.95
)

lm_mverse(.mverse, ...)
}
\arguments{
\item{.mverse}{a \code{mverse} object.}

\item{x}{column name of data within mverse object}

\item{y}{(optional) column name of data within mverse object}

\item{alternative}{a character string specifying the alternative hypothesis,
must be one of "two.sided" (default), "greater" or "less". You can specify just the initial letter.}

\item{mu}{a number indicating the true value of the mean (or difference in means if you are performing a two sample test).}

\item{paired}{a logical indicating whether you want a paired t-test.}

\item{var.equal}{a logical variable indicating whether to treat the two variances as being equal.}

\item{conf.level}{confidence level of the interval.}
}
\value{
A \code{mverse} object with \code{lm} fitted.

A \code{mverse} object with \code{lm} fitted.
}
\description{
\code{t.test.mverse} fits \code{lm} across the multiverse
according to model specifications provided by \code{formula_branch}.
At least one \code{formula_branch} must have been added.

\code{lm_mverse} fits \code{lm} across the multiverse
according to model specifications provided by \code{formula_branch}.
At least one \code{formula_branch} must have been added.
}
\examples{
\dontrun{
model_spec <- formula_branch(y ~ x1)
mv <- mv \%>\%
  add_formula_branch(model_spec) \%>\%
  fit_lm()
}
\dontrun{
model_spec <- formula_branch(y ~ x1)
mv <- mv \%>\%
  add_formula_branch(model_spec) \%>\%
  fit_lm()
}
}
\seealso{
Other {model fitting methods}: 
\code{\link{fit_glm}}

Other {model fitting methods}: 
\code{\link{fit_glm}}
}
\concept{{model fitting methods}}
