% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_branch.R
\name{formula_branch}
\alias{formula_branch}
\title{Create a new formula branch.}
\usage{
formula_branch(..., name = NULL)
}
\arguments{
\item{...}{branch definition expressions.}

\item{name}{Name for the new formula.}
}
\value{
a \code{formula_branch} object.
}
\description{
Create a new formula branch.
}
\examples{
\dontrun{
form <- formula_branch(
  y ~ x1,
  y ~ x2,
  y ~ x1*x2
)
}

}
\seealso{
Other {methods for working with a formula branch}: 
\code{\link{add_formula_branch}()},
\code{\link{family_branch}()}
}
\concept{{methods for working with a formula branch}}
