% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{multiverse_tree}
\alias{multiverse_tree}
\title{Plot a multiverse tree diagram.}
\usage{
multiverse_tree(.mverse, label = FALSE, branches = NULL)
}
\arguments{
\item{.mverse}{A \code{mverse} object.}

\item{label}{A logical. Display options as labels when TRUE.}

\item{branches}{A character vector. Display a subset of branches
when specified. Display all when NULL.}
}
\description{
Plot a multiverse tree diagram.
}
