% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R, R/multiverseExtra.R
\name{create_multiverse}
\alias{create_multiverse}
\alias{multiverseExtra}
\title{Create a new multiverseExtra object}
\usage{
create_multiverse(data)

multiverseExtra(data)
}
\arguments{
\item{data}{Source datafame.}
}
\value{
A multiverseExtra object with the source dataframe attached.
}
\description{
Constructs a new multiverseExtra object which extends
\code{multiverse::multiverse} object.
}
\section{Functions}{
\itemize{
\item \code{create_multiverse}: A descriptive alias
to \code{mutliverseExtra}.
}}

\examples{
\dontrun{
M <- multiverseExtra(df)
}
}
