% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_branch.R
\name{mutate_branch}
\alias{mutate_branch}
\title{Create a new mutate branch.}
\usage{
mutate_branch(..., name = NULL)
}
\arguments{
\item{...}{branch definition expressions.}

\item{name}{Name for the new variable.}
}
\value{
a mutate_branch object.
}
\description{
Create a new mutate branch.
}
\examples{
\dontrun{
x4 <- mutate_branch(x1 + x2, x1 + x3, sum(x1, x2, x3))
}

}
\seealso{
Other {methods for working with a mutate branch}: 
\code{\link{add_mutate_branch}()}
}
\concept{{methods for working with a mutate branch}}
