% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R, R/mverse.R
\name{create_multiverse}
\alias{create_multiverse}
\alias{mverse}
\title{Create a new mverse object}
\usage{
create_multiverse(data)

mverse(data)
}
\arguments{
\item{data}{Source datafame.}
}
\value{
A mverse object with the source dataframe attached.
}
\description{
Constructs a new mverse object which extends
\code{multiverse::multiverse} object.
}
\section{Functions}{
\itemize{
\item \code{create_multiverse}: A descriptive alias
to \code{mverse}.
}}

\examples{
\dontrun{
M <- mverse(df)
}
}
