% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variableBranch.R
\name{name.variableBranch}
\alias{name.variableBranch}
\title{Add a variable name to a variableBranch object.}
\usage{
\method{name}{variableBranch}(vB, name = NULL)
}
\arguments{
\item{vB}{A variableBranch object.}

\item{name}{(optional) Name for the new variable.}
}
\value{
A named variableBranch object if a new \code{name} is provided.
  Otherwise, return the name of the \code{variableBranch} object.
}
\description{
For internal use.
}
