% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch.R
\name{print.branch}
\alias{print.branch}
\title{Print method for \code{*_branch} objects.}
\usage{
\method{print}{branch}(x, ...)
}
\arguments{
\item{x}{a \code{branch} object.}

\item{...}{ignored. for compatibility only.}
}
\description{
Print method for \code{*_branch} objects.
}
