% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_branch.R
\name{remove_filter_branch}
\alias{remove_filter_branch}
\alias{remove_filter_branch.mverse}
\title{Remove a filter branch from a \code{mverse}.}
\usage{
remove_filter_branch(.mverse, name)

\method{remove_filter_branch}{mverse}(.mverse, name)
}
\arguments{
\item{.mverse}{a \code{mverse} object.}

\item{name}{name of the \code{filter branch} to be removed.}
}
\value{
The resulting \code{mverse} object.
}
\description{
This method removes one filter branch from
an existing \code{mverse} object.
}
\examples{
\dontrun{
mv <- create_multiverse(df)
mv <- mverse \%>\%
  remove_filter_branch(old_branch)
}
}
\seealso{
Other {methods for working with a variable branch}: 
\code{\link{remove_mutate_branch}()}
}
\concept{{methods for working with a variable branch}}
