% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_branch.R
\name{remove_mutate_branch}
\alias{remove_mutate_branch}
\alias{remove_mutate_branch.mverse}
\title{Remove a mutate branch from a \code{mverse}.}
\usage{
remove_mutate_branch(.mverse, name)

\method{remove_mutate_branch}{mverse}(.mverse, name)
}
\arguments{
\item{.mverse}{a \code{mverse} object.}

\item{name}{name of the \code{mutate branch} to be removed.}
}
\value{
The resulting \code{mverse} object.
}
\description{
This method removes one mutate branch from
an existing \code{mverse} object.
}
\examples{
\dontrun{
mv <- create_multiverse(df)
mv <- mverse \%>\%
  remove_mutate_branch(old_branch)
}
}
\seealso{
Other {methods for working with a variable branch}: 
\code{\link{remove_filter_branch}()}
}
\concept{{methods for working with a variable branch}}
