% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mverse.R
\name{remove_variable_branch}
\alias{remove_variable_branch}
\alias{remove_variable_branch.mverse}
\title{Remove a variable branch to a \code{multiverse}.}
\usage{
remove_variable_branch(.multiverse, name)

\method{remove_variable_branch}{mverse}(.multiverse, name)
}
\arguments{
\item{.multiverse}{A \code{multiverse} object.}

\item{name}{Variable branch name to be removed.}
}
\description{
This method removes one variable branch from
an existing \code{multiverse} object.
}
\examples{
\dontrun{
mverse <- create_multiverse(df)
mverse <- mverse \%>\%
  remove_variable_branch(old_branch)
}
}
