% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_modules.R
\name{run_module}
\alias{run_module}
\title{Run an interactive module.}
\usage{
run_module(module)
}
\arguments{
\item{module}{module name.}
}
\description{
Run an interactive module.
}
\examples{
run_module(demo)
}
\concept{{interactive modules}}
