% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soccer.R
\docType{data}
\name{soccer}
\alias{soccer}
\title{Number of cards given for each referee-player pair in soccer.}
\format{A data frame with 146,028 rows and 26 variables:
\describe{
  \item{playerShort}{short player ID}
  \item{player}{player name}
  \item{club}{player club}
  \item{leagueCountry}{country of player club (England, Germany, France, and Spain)}
  \item{birthday}{player birthday}
  \item{height}{player height (in cm)}
  \item{weight}{player weight (in kg)}
  \item{position}{detailed player position}
  \item{games}{number of games in the player-referee dyad}
  \item{victories}{victories in the player-referee dyad}
  \item{ties}{ties in the player-referee dyad}
  \item{defeats}{losses in the player-referee dyad}
  \item{goals}{goals scored by a player in the player-referee dyad}
  \item{yellowCards}{number of yellow cards player received from referee}
  \item{yellowReds}{number of yellow-red cards player received from referee}
  \item{redCards}{number of red cards player received from referee}
  \item{rater1}{skin rating of photo by rater 1 (5-point scale ranging from “very light skin” to “very dark skin”)}
  \item{rater2}{skin rating of photo by rater 2 (5-point scale ranging from “very light skin” to “very dark skin”)}
  \item{refNum}{unique referee ID number (referee name removed for anonymizing purposes)}
  \item{refCountry}{unique referee country ID number (country name removed for anonymizing purposes)}
  \item{meanIAT}{mean implicit bias score (using the race IAT) for referee country, higher values correspond to faster white | good, black | bad associations}
  \item{nIAT}{sample size for race IAT in that particular country}
  \item{seIAT}{standard error for mean estimate of race IAT}
  \item{meanExp}{mean explicit bias score (using a racial thermometer task) for referee country, higher values correspond to greater feelings of warmth toward whites versus blacks}
  \item{nExp}{sample size for explicit bias in that particular country}
  \item{seExp}{standard error for mean estimate of explicit bias measure}
}}
\source{
Silberzahn, R., Uhlmann, E. L., Martin, D. P., Anselmi, P., Aust, F., Awtrey, E. C., … Nosek, B. A. (2018, August 24).
\emph{Many analysts, one dataset: Making transparent how variations in analytical choices affect results.}
Retrieved from \url{https://osf.io/gvm2z/}
}
\usage{
soccer
}
\description{
A dataset containing card counts between 2,053 soccer players
and 3,147 referees with other covariates including skin color
ratings. Each line represents a player-referee pair.
}
\details{
The skin colour of each player was rated by two independent raters,
\emph{rater1} and \emph{rater2}, and the 5-point scale values were
scaled to 0 to 1 - i.e., 0, 0.25, 0.5, 0.75, 1.
}
\keyword{datasets}
