% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{summary.glm.nb_mverse}
\alias{summary.glm.nb_mverse}
\title{Display a summary of fitting \code{glm.nb_mverse} across the multiverse.}
\usage{
\method{summary}{glm.nb_mverse}(object, conf.int = TRUE, conf.level = 0.95, output = "estimates", ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Display a summary of fitting \code{glm.nb_mverse} across the multiverse.
}
