% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mverse.R
\name{summary}
\alias{summary}
\alias{summary.mverse}
\title{Display the multiverse table.}
\usage{
summary(.mverse)

\method{summary}{mverse}(.mverse)
}
\arguments{
\item{.mverse}{A \code{mverse} object.}
}
\value{
Multiverse table as a tibble
}
\description{
This method returns the multiverse table
displaying all universes defined by the multiverse.
Each row corresponds to a universe and each column
represents a branch.
}
