% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_branch.R
\name{variable_branch}
\alias{variable_branch}
\title{Create a new variable_branch object}
\usage{
variable_branch(..., name = NULL)
}
\arguments{
\item{...}{Branch definitions.}

\item{name}{(optional) Name for the new variable.}
}
\value{
A variable_branch object.
}
\description{
Create a new variable_branch object
}
\examples{
\dontrun{
x4 <- variableBranch(x1 + x2, x1 + x3, sum(x1, x2, x3))
}

}
