% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variableBranch.R
\name{variableBranch}
\alias{variableBranch}
\title{Create a new variableBranch object}
\usage{
variableBranch(..., name = "")
}
\arguments{
\item{...}{Branch definitions.}

\item{name}{(optional) Name for the new variable.}
}
\value{
A variableBranch object.
}
\description{
Create a new variableBranch object
}
\examples{
\dontrun{
x4 <- variableBranch(x1 + x2, x1 + x3, sum(x1, x2, x3))
}

}
