% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family_branch.R
\name{add_family_branch}
\alias{add_family_branch}
\alias{add_family_branch.mverse}
\title{Add family branches to a \code{mverse} object.}
\usage{
\method{add_family_branch}{mverse}(.mverse, ...)
}
\arguments{
\item{.mverse}{a \code{mverse} object.}

\item{...}{\code{family_branch} objects.}
}
\value{
The resulting \code{mverse} object.
}
\description{
This method adds one or more family branches to
an existing \code{mverse} object. The family branches
are used to specify a model family for the main analysis
that takes a family argument.
}
\examples{
\dontrun{
mv <- create_multiverse(df)
fam <- family_branch(
  poisson, gaussian(link = "log")
)
mv <- mv \%>\%
  add_family_branch(fam)
}
}
\concept{{methods for working with a family branch}}
