% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family_branch.R
\name{add_family_branch}
\alias{add_family_branch}
\alias{add_family_branch.mverse}
\title{Add family branches to a \code{mverse} object.}
\usage{
add_family_branch(.mverse, ...)

\method{add_family_branch}{mverse}(.mverse, ...)
}
\arguments{
\item{.mverse}{a \code{mverse} object.}

\item{...}{\code{family_branch} objects.}
}
\value{
The resulting \code{mverse} object.
}
\description{
This method adds one or more family branches to
an existing \code{mverse} object. The family branches
are used to specify a model family for the main analysis
that takes a family argument.
}
\examples{
# Define a family branch.
model_distributions <- family_branch(
 gaussian, poisson(link = "log")
)
# Create a mverse and add the branch.
mv <- create_multiverse(hurricane) \%>\%
  add_family_branch(model_distributions)
}
\seealso{
Other {methods for working with a family branch}: 
\code{\link{family_branch}()}
}
\concept{{methods for working with a family branch}}
