% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mverse.R
\name{add_variable_branch}
\alias{add_variable_branch}
\alias{add_variable_branch.mverse}
\title{Add variable branches to a \code{multiverse}.}
\usage{
add_variable_branch(.multiverse, ...)

\method{add_variable_branch}{mverse}(.multiverse, ...)
}
\arguments{
\item{.multiverse}{A \code{multiverse} object.}

\item{...}{Variable branch rules.}
}
\description{
This method adds one ore more variable branches to
an existing \code{multiverse} object.
}
\examples{
\dontrun{
mverse <- create_multiverse(df)
new_branch <- variableBranch("x3", x1+x2, mean(c(x1,x2)))
mverse <- mverse \%>\%
  add_variable_branch(new_branch)
}

}
