% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_branch.R
\name{add_variable_branch}
\alias{add_variable_branch}
\alias{add_variable_branch.mverse}
\title{Add variable branches to a \code{mverse} object.}
\usage{
add_variable_branch(.mverse, ...)

\method{add_variable_branch}{mverse}(.mverse, ...)
}
\arguments{
\item{.mverse}{A \code{mverse} object.}

\item{...}{\code{variable_branch} object(s).}
}
\value{
The resulting \code{mverse} object.
}
\description{
This method adds one ore more variable branches to
an existing \code{mverse} object.
}
\examples{
\dontrun{
mv <- create_multiverse(df)
x3 <- variable_branch(x1+x2, mean(c(x1,x2)))
mv <- mv \%>\%
  add_variable_branch(x3)
}
}
