% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{AIC}
\alias{AIC}
\alias{AIC.glm_mverse}
\alias{BIC.glm_mverse}
\alias{BIC}
\title{Display the AIC and BIC score of the fitted models across the multiverse}
\usage{
\method{AIC}{glm_mverse}(object, ..., k = 2)

\method{BIC}{glm_mverse}(object, ...)

AIC(object, ..., k = 2)

BIC(object, ...)
}
\arguments{
\item{object}{a \code{glm_mverse} object.}

\item{...}{ignored. for compatibility only.}

\item{k}{ignored. for compatibility only.}
}
\value{
a multiverse table as a tibble
}
\description{
Display the AIC and BIC score of \code{glm} regression
results across the multiverse.
}
