% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mverse.R
\name{execute_multiverse}
\alias{execute_multiverse}
\alias{execute_multiverse.mverse}
\title{Execute the entire multiverse.}
\usage{
execute_multiverse(.mverse)

\method{execute_multiverse}{mverse}(.mverse)
}
\arguments{
\item{.mverse}{A \code{mverse} object.}
}
\value{
The resulting \code{mverse} object.
}
\description{
This method executes the analysis steps
defined in the \code{mverse} objected
across the entire multiverse.
}
\examples{
\dontrun{
mv <- mv \%>\%
  execute_multiverse(old_branch)
}
}
