% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mverse.R
\name{execute_multiverse}
\alias{execute_multiverse}
\alias{execute_multiverse.mverse}
\title{Execute the entire multiverse.}
\usage{
execute_multiverse(.mverse)

\method{execute_multiverse}{mverse}(.mverse)
}
\arguments{
\item{.mverse}{a \code{mverse} object.}
}
\value{
The resulting \code{mverse} object.
}
\description{
This method executes the analysis steps
defined in the \code{mverse} objected
across the entire multiverse.
}
\examples{
# Create a mverse object
mv <- mverse(hurricane)
# Define and add a mutate branch
femininity <- mutate_branch(
 MasFem > 6, MasFem > mean(MasFem), Gender_MF == 1)
mv <- mv \%>\%
 add_mutate_branch(femininity)
# Execute the multiverse
execute_multiverse(mv)
}
\seealso{
Other {mverse methods}: 
\code{\link{extract}()},
\code{\link{mverse}()}
}
\concept{{mverse methods}}
