% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract}
\alias{extract}
\alias{extract.mverse}
\title{Extract branched values.}
\usage{
extract(...)

\method{extract}{mverse}(
  .mverse,
  columns = NULL,
  universe = NULL,
  nuni = NULL,
  frow = NULL,
  ...
)
}
\arguments{
\item{...}{Ignored.}

\item{.mverse}{a \code{mverse} object.}

\item{columns}{a character vector of column names to extract.}

\item{universe}{an integer vector of universe ids to extract.}

\item{nuni}{a positive integer for the number of universes to extract.}

\item{frow}{proportion of rows to extract from each universe.}
}
\description{
\code{extract} returns a tibble of selected values
across the multiverse in a long format.
}
\details{
This method extracts data values across
the multiverse. You can specify a subset of data
to extract using \code{columns}, \code{universe},
 \code{nuni}, and \code{frow}.

You can specify the columns to extract from each
universe by passing the column names as a character
vector to \code{columns}. The default values is
\code{NULL} extracting all columns with branches.

Use \code{universe} to specify a set of universes
by their integer ids. Use \code{nuni} to specify
the number of universes to extract data from. The
method then selects the subset randomly. Specifying
\code{universe} manually will override \code{nuni} value.
By default, they are both set to \code{NULL} and
the method returns data from all universes.

Use \code{frow} to randomly extract a fraction of
data from each universe. The default value is \code{NULL}
and all rows are returned as they are. Note if you select
1 the method will return shuffle rows in each universe
before returning them. If \code{frow} is greater than 1,
the method randomly samples rows with replacement.
}
\seealso{
Other {mverse methods}: 
\code{\link{execute_multiverse}()},
\code{\link{mverse}()}
}
\concept{{mverse methods}}
