% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family_branch.R
\name{family_branch}
\alias{family_branch}
\title{Create a new family branch.}
\usage{
family_branch(..., name = NULL)
}
\arguments{
\item{...}{branch definition expressions.}

\item{name}{Name for the new family.}
}
\value{
a \code{family_branch} object.
}
\description{
Create a new family branch.
}
\examples{
# Define a family branch.
model_distributions <- family_branch(
  gaussian, poisson(link = "log")
)
# Create a mverse and add the branch.
mv <- create_multiverse(hurricane) \%>\%
  add_family_branch(model_distributions)
}
\seealso{
Other {family branch functions}: 
\code{\link{add_family_branch}()}
}
\concept{{family branch functions}}
