% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{glm.nb_mverse}
\alias{glm.nb_mverse}
\title{Fit a Negative Binomial Regression model across the multiverse
\code{glm.nb_mverse} fits \code{MASS::glm.nb} across the multiverse
according to model specifications provided by \code{formula_branch}.
At least one \code{formula_branch} must have been added.}
\usage{
glm.nb_mverse(.mverse)
}
\description{
Fit a Negative Binomial Regression model across the multiverse
\code{glm.nb_mverse} fits \code{MASS::glm.nb} across the multiverse
according to model specifications provided by \code{formula_branch}.
At least one \code{formula_branch} must have been added.
}
