% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{glm_mverse}
\alias{glm_mverse}
\title{Fit generalized linear regression models across the multiverse.}
\usage{
glm_mverse(.mverse)
}
\arguments{
\item{.mverse}{a \code{mverse} object.}
}
\value{
A \code{mverse} object with \code{glm} fitted.
}
\description{
\code{glm_mverse} fits \code{glm} across the multiverse
according to model specifications provided by \code{formula_branch}.
At least one \code{formula_branch} must have been added.
You can also specify the underlying error distribution and
the link function by adding a \code{family_branch}. If no
\code{family_branch} has been provided, it follows
the default behaviour of \code{glm} using the Gaussian
distribution with an identity link.
}
\examples{
\dontrun{
model_spec <- formula_branch(y ~ x1)
fam <- family_branch(
  poisson, gaussian(link = "log")
)
mv <- mv \%>\%
  add_formula_branch(model_spec) \%>\%
  add_family_banch(fam) \%>\%
  fit_glm()
}
}
