% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{lm_mverse}
\alias{lm_mverse}
\title{Fit linear regression models across the multiverse.}
\usage{
lm_mverse(.mverse)
}
\arguments{
\item{.mverse}{a \code{mverse} object.}
}
\value{
A \code{mverse} object with \code{lm} fitted.
}
\description{
\code{lm_mverse} fits \code{lm} across the multiverse
according to model specifications provided by \code{formula_branch}.
At least one \code{formula_branch} must have been added.
}
\examples{
\dontrun{
model_spec <- formula_branch(y ~ x1)
mv <- mv \%>\%
  add_formula_branch(model_spec) \%>\%
  fit_lm()
}
}
\seealso{
Other {model fitting functions}: 
\code{\link{glm.nb_mverse}()},
\code{\link{glm_mverse}()}
}
\concept{{model fitting functions}}
