% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R, R/multiverseExtra.R
\name{multiverseExtra}
\alias{multiverseExtra}
\alias{create_multiverse}
\title{Create a new multiverseExtra object}
\usage{
create_multiverse(data)

multiverseExtra(data)
}
\arguments{
\item{data}{Source datafame.}
}
\value{
A multiverseExtra object with the source dataframe attached.

A multiverseExtra object with the source dataframe attached.
}
\description{
This is an alias for \code{multiverseExtra} constructor.
Constructs a new multiverseExtra object which extends \code{multiverse::multiverse} object.

Constructs a new multiverseExtra object which extends \code{multiverse::multiverse} object.
}
\examples{
\dontrun{
M <- create_multiverse(df)
}
\dontrun{
M <- multiverseExtra(df)
}
}
