% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mverse.R
\name{mverse}
\alias{mverse}
\alias{create_multiverse}
\title{Create a new \code{mverse} object}
\usage{
mverse(data)

create_multiverse(data)
}
\arguments{
\item{data}{source datafame.}
}
\value{
A \code{mverse} object with the source dataframe attached.
}
\description{
Constructs a new mverse object which extends
\code{multiverse::multiverse} object.
}
\examples{
# Create a mverse object.
mv <- mverse(hurricane)
# create_multiverse() is an alias of mverse().
mv <- create_multiverse(hurricane)
}
