% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_branch.R
\name{name.variable_branch}
\alias{name.variable_branch}
\alias{name}
\title{Internal method for naming a variable branch.}
\usage{
\method{name}{variable_branch}(vB, name = NULL)

name(vB, name = NULL)
}
\description{
Internal method to name the variable
  for a variable branch.
}
