% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_branch.R
\name{remove_variable_branch}
\alias{remove_variable_branch}
\alias{remove_variable_branch.mverse}
\title{Remove a variable branch to a \code{mverse}.}
\usage{
remove_variable_branch(.mverse, name)

\method{remove_variable_branch}{mverse}(.mverse, name)
}
\arguments{
\item{.mverse}{A \code{mverse} object.}

\item{name}{Name of the \code{variable_branch} to be removed.}
}
\value{
The resulting \code{mverse} object.
}
\description{
This method removes one variable branch from
an existing \code{mverse} object.
}
\examples{
\dontrun{
mv <- create_multiverse(df)
mv <- mverse \%>\%
  remove_variable_branch(old_branch)
}
}
