% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_branch.R
\name{remove_variable_branch}
\alias{remove_variable_branch}
\alias{remove_variable_branch.mverse}
\title{Remove a variable branch from a \code{mverse}.}
\usage{
remove_variable_branch(.mverse, name)

\method{remove_variable_branch}{mverse}(.mverse, name)
}
\arguments{
\item{.mverse}{a \code{mverse} object.}

\item{name}{name of the \code{variable_branch} to be removed.}
}
\value{
The resulting \code{mverse} object.
}
\description{
This method removes one variable branch from
an existing \code{mverse} object.
}
\examples{
\dontrun{
mv <- create_multiverse(df)
mv <- mverse \%>\%
  remove_variable_branch(old_branch)
}
}
\seealso{
Other {methods for working with a variable branch}: 
\code{\link{add_variable_branch}()},
\code{\link{variable_branch}()}
}
\concept{{methods for working with a variable branch}}
