% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_curve.R
\name{spec_curve}
\alias{spec_curve}
\alias{spec_curve.lm_mverse}
\alias{spec_curve.glm_mverse}
\title{Display a specification curve across the multiverse.}
\usage{
spec_curve(
  .object,
  var,
  conf.int,
  conf.level,
  option,
  universe_order,
  color_order,
  color,
  branch_order
)

\method{spec_curve}{lm_mverse}(
  .object,
  var = NULL,
  conf.int = TRUE,
  conf.level = 0.95,
  option = names(multiverse::parameters(.object)),
  universe_order = FALSE,
  color_order = FALSE,
  color = "p.value < 0.05",
  branch_order = NULL
)

\method{spec_curve}{glm_mverse}(
  .object,
  var = NULL,
  conf.int = TRUE,
  conf.level = 0.95,
  option = names(multiverse::parameters(.object)),
  universe_order = FALSE,
  color_order = FALSE,
  color = "p.value < 0.05",
  branch_order = NULL
)
}
\arguments{
\item{.object}{a \code{glm_mverse} object.}

\item{var}{name for the variable to show.}

\item{conf.int}{When \code{TRUE} (default), the estimate output
includes the confidence intervals.}

\item{conf.level}{The confidence level of the confidence interval
returned using \code{conf.int = TRUE}. Default value is 0.95.}

\item{option}{a vector of branches to show the options included.}

\item{universe_order}{When \code{TRUE}, order the universes according to
the order in the summary table.}

\item{color_order}{When \code{TRUE}, the estimated value will be ordered
according to the color.}

\item{color}{specify the color in the plot.}

\item{branch_order}{name for the branch to order.}
}
\value{
a specification curve plot for the estimates

a specification curve plot for the estimates
}
\description{
\code{spec_curve.glm_mverse} returns the specification curve of \code{glm}
regression results across the multiverse.
Notice that the order of universes is not corresponding
to the order in the summary table.
}
\examples{
# Create a mverse object
mv <- mverse(hurricane)
# Define and add a mutate branch
femininity <- mutate_branch(
  MasFem > 6, MasFem > mean(MasFem), Gender_MF == 1
)
add_mutate_branch(mv, femininity)
# Define and add a formula branch
model <- formula_branch(
  alldeaths ~ femininity, log(alldeaths + 1) ~ femininity
)
add_formula_branch(mv, model)
# Fit a lm model
lm_mverse(mv)
# Display the specification curve
spec_curve(mv, var = "femininityTRUE")

# Create a mverse object
mv <- mverse(hurricane)
# Define and add a mutate branch
femininity <- mutate_branch(
  MasFem > 6, MasFem > mean(MasFem), Gender_MF == 1
)
add_mutate_branch(mv, femininity)
# Define and add a formula branch
model <- formula_branch(
  alldeaths ~ femininity, alldeaths ~ femininity * HighestWindSpeed
)
add_formula_branch(mv, model)
# Define and add a family branch
model_family <- family_branch(gaussian, poisson)
add_family_branch(mv, model_family)
# Fit a glm model
glm_mverse(mv)
# Display the specification curve
spec_curve(mv, var = "femininityTRUE")

}
\concept{{spec_curve method}}
