% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{summary}
\alias{summary}
\alias{summary.mverse}
\title{Display the multiverse table.}
\usage{
\method{summary}{mverse}(object)
}
\arguments{
\item{object}{a \code{mverse} object.}
}
\value{
a multiverse table as a tibble
}
\description{
This method returns the multiverse table
displaying all universes defined by the multiverse.
Each row corresponds to a universe and each column
represents a branch.
}
\seealso{
Other {mverse methods}: 
\code{\link{execute_multiverse}()},
\code{\link{extract}()},
\code{\link{mverse}()}
}
\concept{{mverse methods}}
