% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{summary}
\alias{summary}
\alias{summary.mverse}
\alias{summary.lm_mverse}
\alias{summary.glm_mverse}
\title{Display the multiverse table with results.}
\usage{
\method{summary}{mverse}(object)

\method{summary}{lm_mverse}(object, conf.int = TRUE, conf.level = 0.95, output = "estimates")

\method{summary}{glm_mverse}(object, conf.int = TRUE, conf.level = 0.95, output = "estimates")
}
\arguments{
\item{object}{a \code{mverse} object.}

\item{conf.int}{When \code{TRUE} (default), the estimate output
includes the confidence intervals.}

\item{conf.level}{The confidence level of the confindence interval
returned using \code{conf.int = TRUE}. Default value is 0.95.}

\item{output}{The output of interest. The possible values are
"estimates" ("e"), "df", "deviance" ("de"), and "aic" ("bic").
Alternatively, the first letters may be used. Default value
is "estimates".}
}
\value{
a multiverse table as a tibble.

a multiverse table as a tibble

a multiverse table as a tibble
}
\description{
This method returns the multiverse table
displaying all universes defined by the multiverse.
Each row corresponds to a universe and each column
represents a branch.

\code{summary.lm_mverse} returns the \code{lm} regression
results across the multiverse.

\code{summary.lm_mverse} returns the \code{lm} regression
results across the multiverse.
}
\concept{{summary method}}
