% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{summary}
\alias{summary}
\alias{summary.mverse}
\alias{summary.lm_mverse}
\title{Display the multiverse table with results.}
\usage{
\method{summary}{mverse}(object)

\method{summary}{lm_mverse}(object, conf.int = FALSE, conf.level = 0.95)
}
\arguments{
\item{object}{a \code{mverse} object.}
}
\value{
a multiverse table as a tibble

a multiverse table as a tibble
}
\description{
This method returns the multiverse table
displaying all universes defined by the multiverse.
Each row corresponds to a universe and each column
represents a branch.

\code{summary.lm_mverse} returns the multiverse table
displaying all universes defined by the multiverse.
Each row corresponds to a universe and each column
represents a branch.
}
\concept{{summary method}}
