% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{summary}
\alias{summary}
\alias{summary.mverse}
\alias{summary.lm_mverse}
\alias{summary.glm_mverse}
\title{Display the multiverse table with results.}
\usage{
\method{summary}{mverse}(object, ...)

\method{summary}{lm_mverse}(object, conf.int = TRUE, conf.level = 0.95, output = "estimates", ...)

\method{summary}{glm_mverse}(object, conf.int = TRUE, conf.level = 0.95, output = "estimates", ...)
}
\arguments{
\item{object}{a \code{glm_mverse} object.}

\item{...}{Ignored.}

\item{conf.int}{When \code{TRUE} (default), the estimate output
includes the confidence intervals.}

\item{conf.level}{The confidence level of the confindence interval
returned using \code{conf.int = TRUE}. Default value is 0.95.}

\item{output}{The output of interest. The possible values are
"estimates" ("e"), "df", "deviance" ("de"), and "aic" ("bic").
Alternatively, the first letters may be used. Default value
is "estimates".}
}
\value{
a multiverse table as a tibble.

a multiverse table as a tibble

a multiverse table as a tibble
}
\description{
This method returns the multiverse table
displaying all universes defined by the multiverse.
Each row corresponds to a universe and each column
represents a branch.

\code{summary.lm_mverse} returns the \code{lm} regression
results across the multiverse.

\code{summary.lm_mverse} returns the \code{glm} regression
results across the multiverse.
}
\examples{
# Define a mutate branch.
hurricane_strength <- mutate_branch(
  # damage vs. wind speed vs.pressure
  NDAM,
  HighestWindSpeed,
  Minpressure_Updated_2014,
  # Standardized versions
  scale(NDAM),
  scale(HighestWindSpeed),
  -scale(Minpressure_Updated_2014),
)
# Create a mverse and add the branch.
mv <- create_multiverse(hurricane) \%>\%
  add_mutate_branch(hurricane_strength)
# Display the multiverse table while
# defining a multiverse provides a way
# to inspect the universes created.
summary(mv)
# Define a filter branch.
hurricane_outliers <- filter_branch(
  ! Name \%in\% c("Katrina", "Audrey", "Andrew"),
  ! Name \%in\% c("Katrina"),
  ! Name \%in\% c("Katrina"),
  TRUE # include all
)
# Add the additional branch and inspect.
mv <- add_filter_branch(mv, hurricane_outliers)
summary(mv)
# Define mutate branches.
hurricane_strength <- mutate_branch(
# damage vs. wind speed vs.pressure
NDAM,
HighestWindSpeed,
Minpressure_Updated_2014,
# Standardized versions
scale(NDAM),
scale(HighestWindSpeed),
-scale(Minpressure_Updated_2014),
)
y <- mutate_branch(
alldeaths, log(alldeaths + 1)
)
# Define a filter branch.
hurricane_outliers <- filter_branch(
! Name \%in\% c("Katrina", "Audrey", "Andrew"),
! Name \%in\% c("Katrina"),
! Name \%in\% c("Katrina"),
TRUE # include all
)
# Define a formula branch.
model_specifications <- formula_branch(
y ~ femininity,
y ~ femininity + hurricane_strength,
y ~ femininity * hurricane_strength
)
# Create a mverse, add the branches, and fit lm models.
mv <- create_multiverse(hurricane) \%>\%
add_filter_branch(hurricane_outliers) \%>\%
add_mutate_branch(hurricane_strength, y) \%>\%
add_formula_branch(model_specifications) \%>\%
lm_mverse()
# Display the multiverse table with estimated coefficients.
summary(mv)
# Define mutate branches.
hurricane_strength <- mutate_branch(
# damage vs. wind speed vs.pressure
NDAM,
HighestWindSpeed,
Minpressure_Updated_2014,
# Standardized versions
scale(NDAM),
scale(HighestWindSpeed),
-scale(Minpressure_Updated_2014),
)
y <- mutate_branch(
alldeaths, log(alldeaths + 1)
)
# Define a filter branch.
hurricane_outliers <- filter_branch(
  ! Name \%in\% c("Katrina", "Audrey", "Andrew"),
! Name \%in\% c("Katrina"),
! Name \%in\% c("Katrina"),
TRUE # include all
)
# Define a formula branch.
model_specifications <- formula_branch(
y ~ femininity,
y ~ femininity + hurricane_strength,
y ~ femininity * hurricane_strength
)
# Define a family branch.
model_distributions <- family_branch(gaussian, poisson)
# Create a mverse, add the branches, and fit glm models.
mv <- create_multiverse(hurricane) \%>\%
add_filter_branch(hurricane_outliers) \%>\%
add_mutate_branch(hurricane_strength, y) \%>\%
add_formula_branch(model_specifications) \%>\%
add_family_branch(model_distributions) \%>\%
glm_mverse()
# Display the multiverse table with estimated coefficients.
summary(mv)
}
\concept{{summary method}}
