% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_branch.R
\name{variable_branch}
\alias{variable_branch}
\title{Create a new variable branch.}
\usage{
variable_branch(..., name = NULL)
}
\arguments{
\item{...}{branch definition expressions.}

\item{name}{Name for the new variable.}
}
\value{
a variable_branch object.
}
\description{
Create a new variable branch.
}
\examples{
\dontrun{
x4 <- variableBranch(x1 + x2, x1 + x3, sum(x1, x2, x3))
}

}
\seealso{
Other {methods for working with a variable branch}: 
\code{\link{add_variable_branch}()},
\code{\link{remove_variable_branch}()}
}
\concept{{methods for working with a variable branch}}
