% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_branch.R
\name{variable_branch}
\alias{variable_branch}
\alias{parse}
\alias{parse.variable_branch}
\title{Create a new variableBranch object}
\usage{
variable_branch(..., name = NULL)

parse(rules, name)

\method{parse}{variable_branch}(vB)
}
\arguments{
\item{...}{Branch definitions.}

\item{name}{(optional) Name for the new variable.}
}
\value{
A variableBranch object.
}
\description{
Create a new variableBranch object
}
\examples{
\dontrun{
x4 <- variableBranch(x1 + x2, x1 + x3, sum(x1, x2, x3))
}

}
